# docs/swagger.yaml
basePath: /api/v2/
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.UpdateCryptoUserApproveRequest:
    properties:
      approve:
        type: boolean
      uuid:
        type: string
    type: object
  handler.UpdateDisplayCodeRequest:
    properties:
      crypto_user_id:
        type: string
      display_code:
        type: string
    type: object
...
  model.CRMLoginBody:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  model.KOLUser:
    properties:
      displayCode:
        type: string
      id:
        type: string
    type: object
...
info:
  contact:
    email: admintool@investicstudio.com
    name: API Support
  description: API documentation for Crypto Studio Backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Crypto Studio Backend API
  version: "1.0"
paths:
...
  /crm/get-crypto-user:
    get:
      consumes:
      - application/json
      description: Get crypto user data
      parameters:
      - default: 1
        description: Page number (starting from 1)
        in: query
        name: page
        type: integer
      - default: desc
        description: Sort by waiting_list_timestamp (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: Search by twitter_name or email
        in: query
        name: search
        type: string
      - description: Filter by copytrade approval (true/false)
        in: query
        name: is_copytrade_approved
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get crypto user data
      tags:
      - CRM
  /crm/get_refferal_score:
    get:
      consumes:
      - application/json
      description: Get refferal score
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get refferal score
      tags:
      - CRM
  /crm/get_user_referral:
    get:
      consumes:
      - application/json
      description: Get User Referral
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get User Referral
      tags:
      - CRM
  /crm/get_users:
    get:
      consumes:
      - application/json
      description: Returns all users with uuid, email, and twitter_uid
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - CRM
  /crm/kol_refer_detail:
    get:
      consumes:
      - application/json
      description: Fetches KOL referral details with pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get KOL referral details
      tags:
      - CRM
  /crm/login:
    post:
      consumes:
      - application/json
      description: Authenticates CRM user and returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CRMLoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: 'Successful login, returns JWT token. Example: {\"token\":
            \"<jwt>\"}'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "400":
          description: Bad Request - Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized - Invalid credentials
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: CRM User login
      tags:
      - CRM
  /crm/new-kol:
    post:
      consumes:
      - application/json
      description: Creates a new KOL user and associates a KOL code. Requires valid
        CRM user and unique display code.
      parameters:
      - description: KOL User Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.KOLUser'
      produces:
      - application/json
      responses:
        "200":
          description: 'KOL user created successfully", e.g., {\"kolUserID\": \"abc123\"}'
        "400":
          description: Bad Request - Invalid input, validation failed, or duplicate
            KOL user
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found - CRM user does not exist
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error - Validation or DB operation failed
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new KOL user
      tags:
      - CRM
  /crm/privy-user-overview:
    get:
      consumes:
      - application/json
      description: Returns user profile, wallets (with subscriptions), and trade logs.
        Filter logs by status and order by execution date.
      parameters:
      - description: User ID (crypto_user.id)
        in: query
        name: user_id
        required: true
        type: string
      - description: Filter trade_logs.status (e.g. success, fail)
        in: query
        name: status
        type: string
      - default: desc
        description: Order by executed_at (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - default: 200
        description: Max logs to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Logs offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search privy copytrade user overview by user id
      tags:
      - CRM
  /crm/update-crypto-user-approve:
    post:
      consumes:
      - application/json
      description: Update a user's copytrade approval (waiting list) status
      parameters:
      - description: Approve payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCryptoUserApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update copytrade approval status
      tags:
      - CRM
  /crm/update_display_code:
    post:
      consumes:
      - application/json
      description: Updates the display_code for a user if not duplicated
      parameters:
      - description: Update display_code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDisplayCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update display_code for a user
      tags:
      - CRM
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your-token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"